grafana-otel-example/product-info-querier/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
deprecated==1.2.14 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.5 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.28.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-common==1.28.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-http==1.28.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-logging==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-redis==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation-threading==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-instrumentation==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-proto==1.28.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-sdk==1.28.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-semantic-conventions==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-util-http==0.49b1 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
redis==5.2.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.41.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.32.0 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
